refactor: BottomAlignedDialogとLabeledIconButtonを共通パッケージに移動
BottomAlignedDialogとLabeledIconButtonコンポーネントをui/utilからui/commonパッケージに移動しました。
また、ThreadToolbarOverflowMenuのレイアウト調整、及びDisplaySettingsアイコンをTuneアイコンに変更しました。
