+++ app/src/main/java/com/websarva/wings/android/slevo/ui/thread/dialog/ThreadToolbarOverflowMenu.kt
@@ -7,13 +7,12 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.ui.Alignment
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.List
 import androidx.compose.material.icons.filled.History
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.filled.Star
-import androidx.compose.material.icons.filled.DisplaySettings
+import androidx.compose.material.icons.filled.Tune
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -21,8 +20,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.websarva.wings.android.slevo.R
-import com.websarva.wings.android.slevo.ui.util.BottomAlignedDialog
-import com.websarva.wings.android.slevo.ui.util.LabeledIconButton
+import com.websarva.wings.android.slevo.ui.common.BottomAlignedDialog
+import com.websarva.wings.android.slevo.ui.common.LabeledIconButton
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -58,13 +57,17 @@ fun ThreadToolbarMenuContent(
     Column(
         modifier = Modifier
             .fillMaxWidth()
-            .padding(vertical = 24.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
+            .padding(vertical = 16.dp, horizontal = 16.dp),
     ) {
         Row(
             modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceAround
+            horizontalArrangement = Arrangement.SpaceBetween
         ) {
+            LabeledIconButton(
+                icon = Icons.Filled.Tune,
+                label = stringResource(R.string.display_settings),
+                onClick = onDisplaySettingsClick
+            )
             LabeledIconButton(
                 icon = Icons.Filled.Star,
                 label = stringResource(R.string.bookmark),
@@ -80,17 +83,12 @@ fun ThreadToolbarMenuContent(
                 label = stringResource(R.string.history),
                 onClick = onHistoryClick
             )
-            LabeledIconButton(
-                icon = Icons.Filled.Settings,
-                label = stringResource(R.string.settings),
-                onClick = onSettingsClick
-            )
         }
-        Spacer(modifier = Modifier.height(24.dp))
+        Spacer(modifier = Modifier.height(8.dp))
         LabeledIconButton(
-            icon = Icons.Filled.DisplaySettings,
-            label = stringResource(R.string.display_settings),
-            onClick = onDisplaySettingsClick
+            icon = Icons.Filled.Settings,
+            label = stringResource(R.string.settings),
+            onClick = onSettingsClick
         )
     }
 }
